Further split apart resources routes into retrieval and update modules
