+++ app/api/routes/__init__.py
@@ -1,6 +1,7 @@
 from app.api.routes import api_key
-from app.api.routes import resources
-from app.api.routes import resource_voting
 from app.api.routes import categories
 from app.api.routes import languages
+from app.api.routes import resource_retrieval
+from app.api.routes import resource_updating
+from app.api.routes import resource_voting
 from app.api.routes import searching
